requirements.txt
Flask
Flask-Mail
Flask-SQLAlchemy
Werkzeug
python-dotenv
pdfkit
crewai
openai
requests
python-docx
